backend/requirements.txt
fastapi>=0.114.0
uvicorn[standard]>=0.30.0
openai>=1.51.0
chromadb>=0.5.0
python-dotenv>=1.0.1